+++ pkg/storegateway/parquet_bucket_stores.go
@@ -23,6 +23,7 @@ import (
 	storecache "github.com/thanos-io/thanos/pkg/store/cache"
 	"github.com/thanos-io/thanos/pkg/store/storepb"
 	"github.com/weaveworks/common/httpgrpc"
+	"github.com/weaveworks/common/user"
 	"golang.org/x/sync/errgroup"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
@@ -56,6 +57,7 @@ type ParquetBucketStores struct {
 
 	matcherCache      storecache.MatchersCache
 	parquetShardCache parquetutil.CacheInterface[parquet_storage.ParquetShard]
+	rowRangesCache    search.RowRangesForConstraintsCache
 
 	inflightRequests *cortex_util.InflightRequestTracker
 }
@@ -73,6 +75,15 @@ func newParquetBucketStores(cfg tsdb.BlocksStorageConfig, bucketClient objstore.
 		return nil, err
 	}
 
+	rowRangesCacheBackend, err := tsdb.CreateParquetRowRangesCache(cfg.BucketStore.ParquetRowRangesCache, logger, reg)
+	if err != nil {
+		return nil, err
+	}
+	var rowRangesCache search.RowRangesForConstraintsCache
+	if rowRangesCacheBackend != nil {
+		rowRangesCache = parquetutil.NewRowRangesCache(rowRangesCacheBackend, "parquet-row-ranges", cfg.BucketStore.ParquetRowRangesCache.TTL, reg)
+	}
+
 	u := &ParquetBucketStores{
 		logger:            logger,
 		cfg:               cfg,
@@ -83,6 +94,7 @@ func newParquetBucketStores(cfg tsdb.BlocksStorageConfig, bucketClient objstore.
 		chunksDecoder:     schema.NewPrometheusParquetChunksDecoder(chunkenc.NewPool()),
 		inflightRequests:  cortex_util.NewInflightRequestTracker(),
 		parquetShardCache: parquetShardCache,
+		rowRangesCache:    rowRangesCache,
 	}
 
 	if cfg.BucketStore.MatchersCacheMaxItems > 0 {
@@ -109,6 +121,7 @@ func (u *ParquetBucketStores) Series(req *storepb.SeriesRequest, srv storepb.Sto
 		return fmt.Errorf("no userID")
 	}
 
+	spanCtx = user.InjectOrgID(spanCtx, userID)
 	err := u.getStoreError(userID)
 	userBkt := bucket.NewUserBucketClient(userID, u.bucket, u.limits)
 	if err != nil {
@@ -149,6 +162,7 @@ func (u *ParquetBucketStores) LabelNames(ctx context.Context, req *storepb.Label
 	if userID == "" {
 		return nil, fmt.Errorf("no userID")
 	}
+	spanCtx = user.InjectOrgID(spanCtx, userID)
 
 	err := u.getStoreError(userID)
 	userBkt := bucket.NewUserBucketClient(userID, u.bucket, u.limits)
@@ -165,7 +179,7 @@ func (u *ParquetBucketStores) LabelNames(ctx context.Context, req *storepb.Label
 		return nil, status.Error(codes.Internal, err.Error())
 	}
 
-	return store.LabelNames(ctx, req)
+	return store.LabelNames(spanCtx, req)
 }
 
 // LabelValues implements BucketStores
@@ -177,6 +191,7 @@ func (u *ParquetBucketStores) LabelValues(ctx context.Context, req *storepb.Labe
 	if userID == "" {
 		return nil, fmt.Errorf("no userID")
 	}
+	spanCtx = user.InjectOrgID(spanCtx, userID)
 
 	err := u.getStoreError(userID)
 	userBkt := bucket.NewUserBucketClient(userID, u.bucket, u.limits)
@@ -193,7 +208,7 @@ func (u *ParquetBucketStores) LabelValues(ctx context.Context, req *storepb.Labe
 		return nil, status.Error(codes.Internal, err.Error())
 	}
 
-	return store.LabelValues(ctx, req)
+	return store.LabelValues(spanCtx, req)
 }
 
 // SyncBlocks implements BucketStores
@@ -262,19 +277,21 @@ func (u *ParquetBucketStores) createParquetBucketStore(userID string, userLogger
 		chunksDecoder:     u.chunksDecoder,
 		matcherCache:      u.matcherCache,
 		parquetShardCache: u.parquetShardCache,
+		rowRangesCache:    u.rowRangesCache,
 	}
 
 	return store, nil
 }
 
 type parquetBlock struct {
-	name        string
-	shard       parquet_storage.ParquetShard
-	m           *search.Materializer
-	concurrency int
+	name           string
+	shard          parquet_storage.ParquetShard
+	m              *search.Materializer
+	concurrency    int
+	rowRangesCache search.RowRangesForConstraintsCache
 }
 
-func (p *parquetBucketStore) newParquetBlock(ctx context.Context, name string, shardID int, labelsFileOpener, chunksFileOpener parquet_storage.ParquetOpener, d *schema.PrometheusParquetChunksDecoder, rowCountQuota *search.Quota, chunkBytesQuota *search.Quota, dataBytesQuota *search.Quota) (*parquetBlock, error) {
+func (p *parquetBucketStore) newParquetBlock(ctx context.Context, name string, shardID int, labelsFileOpener, chunksFileOpener parquet_storage.ParquetOpener, d *schema.PrometheusParquetChunksDecoder, rowRangesCache search.RowRangesForConstraintsCache, rowCountQuota *search.Quota, chunkBytesQuota *search.Quota, dataBytesQuota *search.Quota) (*parquetBlock, error) {
 	userID, err := users.TenantID(ctx)
 	if err != nil {
 		return nil, err
@@ -316,10 +333,11 @@ func (p *parquetBucketStore) newParquetBlock(ctx context.Context, name string, s
 	}
 
 	return &parquetBlock{
-		shard:       shard,
-		m:           m,
-		concurrency: p.concurrency,
-		name:        name,
+		shard:          shard,
+		m:              m,
+		concurrency:    p.concurrency,
+		name:           name,
+		rowRangesCache: rowRangesCache,
 	}, nil
 }
 
@@ -385,8 +403,8 @@ func (b *parquetBlock) Query(ctx context.Context, mint, maxt int64, skipChunks b
 			if err != nil {
 				return err
 			}
-			// TODO: Add cache.
-			rr, err := search.Filter(ctx, b.shard, rgi, nil, cs...)
+
+			rr, err := search.Filter(ctx, b.shard, rgi, b.rowRangesCache, cs...)
 			if err != nil {
 				return err
 			}
@@ -446,8 +464,8 @@ func (b *parquetBlock) LabelNames(ctx context.Context, limit int64, matchers []*
 			if err != nil {
 				return err
 			}
-			// TODO: Add cache.
-			rr, err := search.Filter(ctx, b.shard, rgi, nil, cs...)
+
+			rr, err := search.Filter(ctx, b.shard, rgi, b.rowRangesCache, cs...)
 			if err != nil {
 				return err
 			}
@@ -487,8 +505,8 @@ func (b *parquetBlock) LabelValues(ctx context.Context, name string, limit int64
 			if err != nil {
 				return err
 			}
-			// TODO: Add cache.
-			rr, err := search.Filter(ctx, b.shard, rgi, nil, cs...)
+
+			rr, err := search.Filter(ctx, b.shard, rgi, b.rowRangesCache, cs...)
 			if err != nil {
 				return err
 			}